Refactor hooks and fix blur() after pattern removal
